Fixed issue pointed out by static analysis. The vast majority of issues is related to wider types but arithmetic expressions (address calculation) is performed in original/lower type-width.

+++ include/libxs_frontend.h
@@ -274,7 +274,7 @@
   LIBXS_MMCALL_ABC(FN, A, B, C)
 #else
 # define LIBXS_MMCALL_LDX(FN, A, B, C, M, N, K, LDA, LDB, LDC) \
-  LIBXS_MMCALL_PRF(FN, A, B, C, (A) + (LDA) * (K), (B) + (LDB) * (N), (C) + (LDC) * (N))
+  LIBXS_MMCALL_PRF(FN, A, B, C, (A) + ((size_t)LDA) * (K), (B) + ((size_t)LDB) * (N), (C) + ((size_t)LDC) * (N))
 #endif
 #define LIBXS_MMCALL(FN, A, B, C, M, N, K) LIBXS_MMCALL_LDX(FN, A, B, C, M, N, K, M, K, M)
